refactor messages and media related apis
