+++ src/components/Compose/Compose.vue
@@ -52,7 +52,7 @@ export default {
 
             // send image, if one is attached
             if (this.$store.state.loaded_media) {
-                Api.sendFile(this.$store.state.loaded_media, (file, messageId) => {
+                Api.messages.media.send(this.$store.state.loaded_media, (file, messageId) => {
                     Api.createThreadWithImage(to, messageId, file.type);
                 });
             }
