+++ src/utils/api/index.js
@@ -1,9 +1,11 @@
-import Stream from '@/utils/api/stream.js'
-import Account from '@/utils/api/account.js'
-import Conversations from '@/utils/api/conversations.js'
+import Stream from '@/utils/api/stream.js';
+import Account from '@/utils/api/account.js';
+import Conversations from '@/utils/api/conversations.js';
+import Messages from '@/utils/api/messages.js';
 
 export {
     Account,
     Conversations,
+    Messages,
     Stream
 }
